solve the compile problem of naxsi

+++ 3rd-party/naxsi/naxsi_src/naxsi_runtime.c
@@ -446,7 +446,7 @@ ngx_http_dummy_is_rule_whitelisted_rx(ngx_http_request_t *req,
 		  "%d/%d RXMZ rule has %d custom locations", i, cf->rxmz_wlr->nelts, 
 		  p->br->custom_locations->nelts);
 #endif
-    if (p->br->zone != zone) {
+    if ((unsigned int)p->br->zone != (unsigned int)zone) {
 #ifdef wlrx_debug
     ngx_log_debug(NGX_LOG_DEBUG_HTTP, req->connection->log, 0, 
 		  "%d/%d Not targeting same zone.");
